:recycle: Factor out file queries
Moves file queries into their own file within the files/schema/
directory.
