+++ creator/schema.py
@@ -12,14 +12,14 @@ class Query(creator.files.schema.Query,
 
 
 class Mutation(graphene.ObjectType):
-    create_file = creator.files.schema.UploadMutation.Field(
+    create_file = creator.files.schema.file.UploadMutation.Field(
         description=(
             "Upload a new file or version to an existing file and "
             "study"
         )
     )
-    update_file = creator.files.schema.FileMutation.Field()
-    delete_file = creator.files.schema.DeleteFileMutation.Field()
+    update_file = creator.files.schema.file.FileMutation.Field()
+    delete_file = creator.files.schema.file.DeleteFileMutation.Field()
     update_version = creator.files.schema.version.VersionMutation.Field()
     signed_url = creator.files.schema.SignedUrlMutation.Field()
     create_dev_token = creator.files.schema.DevDownloadTokenMutation.Field()
